Develop

+++ TKDprogress_BLL/Services/CategoryService.cs
@@ -3,9 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TKDprogress_BLL.Models;
 using TKDprogress_BLL.Interfaces;
-using TKDprogress_SL.Entities;
-using TKDprogress_SL.Interfaces;
+using TKDprogress_BLL.Interfaces.Repositories;
+using TKDprogress_BLL.Interfaces.Services;
 
 namespace TKDprogress_BLL.Services
 {
@@ -18,36 +19,70 @@ namespace TKDprogress_BLL.Services
             _categoryRepository = categoryRepository;
         }
 
-        public async Task<List<CategoryDto>> GetCategoriesAsync(string searchString)
+        public async Task<List<Category>> GetCategoriesAsync(string searchString)
         {
-            List<CategoryDto> categories = await _categoryRepository.GetCategoriesAsync(searchString);
+            List<Category> categories = await _categoryRepository.GetCategoriesAsync(searchString);
 
             return categories;
         }
 
-        public async Task<CategoryDto> GetCategoryByIdAsync(int id)
+        public async Task<Category> GetCategoryByIdAsync(int id)
         {
-            CategoryDto artist = await _categoryRepository.GetCategoryByIdAsync(id);
+            if (id <= 0)
+            {
+                return new Category { ErrorMessage = "Invalid category." };
+            }
 
-            return artist;
-        }
+            Category category = await _categoryRepository.GetCategoryByIdAsync(id);
 
-        public async Task<CategoryDto> CreateCategoryAsync(CategoryDto category)
-        {
-            await _categoryRepository.CreateCategoryAsync(category);
+            if (category.Id <= 0)
+            {
+                return new Category { ErrorMessage = "Category not found." };
+            }
 
             return category;
         }
 
-        public async Task<CategoryDto> UpdateCategoryAsync(CategoryDto category)
+        public async Task<Category> CreateCategoryAsync(Category category)
         {
-            await _categoryRepository.UpdateCategoryAsync(category);
+            if (string.IsNullOrEmpty(category.Name) || string.IsNullOrEmpty(category.Description))
+            {
+                category.ErrorMessage =  "Category name or description has an incorrect input.";
+                return category;
+            }
+
+            category = await _categoryRepository.CreateCategoryAsync(category);
 
             return category;
         }
 
-        public async Task<CategoryDto> DeleteCategoryAsync(CategoryDto category)
+        public async Task<Category> UpdateCategoryAsync(Category category)
         {
+            if (category.Id <= 0)
+            {
+                category.ErrorMessage = "Category type does not exist.";
+                return category;
+            }
+
+            if (string.IsNullOrEmpty(category.Name) || string.IsNullOrEmpty(category.Description))
+            {
+                category.ErrorMessage = "Category name or description has an incorrect input.";
+                return category;
+            }
+
+            category = await _categoryRepository.UpdateCategoryAsync(category);
+
+            return category;
+        }
+
+        public async Task<Category> DeleteCategoryAsync(Category category)
+        {
+            if (category.Id <= 0)
+            {
+                category.ErrorMessage = "Category type does not exist.";
+                return category;
+            }
+
             await _categoryRepository.DeleteCategoryAsync(category);
 
             return category;
